Treat datasource ID as an integer

+++ pkg/grafana/datasources.go
@@ -96,7 +96,7 @@ func postDatasource(resource grizzly.Resource) error {
 
 func putDatasource(resource grizzly.Resource) error {
 	spec := resource.Spec()
-	id := spec["id"].(int64)
+	id := int64(spec["id"].(float64))
 	grafanaURL, err := getGrafanaURL(fmt.Sprintf("api/datasources/%d", id))
 	if err != nil {
 		return err
